+++ src/main/java/com/victor/portfolio/restapi/product/dto/ProductResponse.java
@@ -6,6 +6,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -16,13 +18,12 @@ public class ProductResponse {
     private Long id;
     private String sku;
     private String name;
-    private Double price;
+    private BigDecimal price;
 
     private Boolean active;
 
     private Long categoryId;
 
-    // 👇 ESTOS DOS CAMPOS SON LOS QUE FALTABAN
     private String categoryCode;
     private String categoryName;
 }
